OpenAPI 3.1 component (#904)

+++ src/Component/OpenApi3/Tests/MultipartNestedObjectRuntimeTest.php
@@ -36,8 +36,9 @@ class MultipartNestedObjectRuntimeTest extends TestCase
 
         $body = new Expected\Model\FilePostBody();
         $body->setFichier('file-content');
-        /** @phpstan-ignore-next-line */
-        $body->setItem($item);
+        if (method_exists($body, 'setItem')) {
+            $body->setItem($item);
+        }
 
         $normalizers = [
             new \Symfony\Component\Serializer\Normalizer\ArrayDenormalizer(),
